Add products and budgets.

+++ AppCalisto/Controllers/ClientsController.cs
@@ -186,10 +186,17 @@ namespace AppCalisto.Controllers
                 return RedirectToAction("NotFound404", "Errors", new { entityName = "Client" });
             }
 
+            var existingClients = await _clientRepository.GetAllAsync();
+            if (existingClients.Count == 1)
+            {
+                TempData["Failure"] = "There must be at least one client in the database.";
+                return RedirectToAction("Edit", new { id });
+            }
+
             if (!await _clientRepository.DeleteAsync(client))
             {
                 TempData["Failure"] = "Could not delete client. This may be due to the entity being used by other entities or no longer existing.";
-                return RedirectToAction("Edit", new { id = client.Id });
+                return RedirectToAction("Edit", new { id });
             }
 
             TempData["Success"] = "Client deleted successfully.";
